Update paru for use with pandoc 2.18
compared to pandoc 2.17, 2.18 does not bring breaking changes. I've
updated paru to reflect the small changes in the order of command line
options as listed in the manual, as well as the change of output for one
test. As a result, no new release of paru needs to be created.

I also updated the Dockerfile to use Ruby 3.1 and pandoc 2.18.1.