Add before and after methods to a filter (#38)
I've added a before and after method to paru pandoc filters. See the
[example filter](examples/filters/before_during_after.rb). The block to
the `before` method is run once, just before the document is filtered.
Similarly, the block to the `after` method is run once, just after the
document is filtered.

I've also added the pseudo selector "*", denoting any node. By default,
if you put code in a filter outside the `with` methods, that code is run
for each and every node in the document you're filtering. That's quite
unexpected behavior. But it is also behavior I cannot change given the
implementation of filters in paru.

With the introduction of the any selector and the before and after
methods, I hope to entice users to stop writing code outside of the
`before`, `with` or `after` methods in their filters. If they want to
run something once, I recommend the `before` method. If they really want
to run code for each node in a document, they can do this explicitly by
using the any selector. However, to ensure backwards compatability, the
old behavior has not changed.